+++ src/firewallfabrik/platforms/iptables/_policy_compiler.py
@@ -45,6 +45,7 @@ from firewallfabrik.compiler.processors._generic import (
     ResolveMultiAddress,
     SimplePrintProgress,
     SingleRuleFilter,
+    VerifyAddresses,
     VerifyAddressRanges,
     VerifyMacAddresses,
 )
@@ -533,6 +534,7 @@ class PolicyCompiler_ipt(PolicyCompiler):
         self.add(VerifyCustomServices('verify custom services'))
         self.add(VerifyPortRanges('verify port ranges'))
         self.add(VerifyAddressRanges('verify address ranges'))
+        self.add(VerifyAddresses('verify addresses'))
         self.add(VerifyMacAddresses('verify MAC addresses'))
         self.add(SpecialCasesWithCustomServices('special cases with custom services'))
         self.add(SeparatePortRanges('separate port ranges'))