+++ src/firewallfabrik/platforms/nftables/_nat_compiler.py
@@ -39,6 +39,7 @@ from firewallfabrik.compiler.processors._generic import (
     RecursiveGroupsInRE,
     ResolveMultiAddress,
     SimplePrintProgress,
+    VerifyAddresses,
     VerifyAddressRanges,
     VerifyMacAddresses,
     VerifyRules,
@@ -318,6 +319,7 @@ class NATCompiler_nft(NATCompiler):
         self.add(VerifyCustomServices('verify custom services'))
         self.add(VerifyPortRanges('verify port ranges'))
         self.add(VerifyAddressRanges('verify address ranges'))
+        self.add(VerifyAddresses('verify addresses'))
         self.add(VerifyMacAddresses('verify MAC addresses'))
         self.add(VerifyRules2('check correctness of TSrv'))
         self.add(SeparatePortRanges('separate port ranges'))